Refactor, implement protection against infinite recursion
It is trivial to configure a domain in the DNS system to
be cross referential with another domain, which in prior
to this commit would cause a panic due to running out of
stack space due to infinite recursion. This code introduces
a test case for this and implements two mechanisms to protect
from this problem.
* Each query now keeps track of seen queries, and if the
  same name and record type pair shows up twice in the same
  lookup, it will fail as this is a sign of two cross referencing
  domains.
* As a back stop, there is also a limit of 5 levels of recursion
  implemented.